Changes: separated tests for hooks and options. Added test specs to check consistency between available options in Handsontable and supported @inputs in the component.
