+++ projects/hot-table/src/lib/hot-table.component.ts
@@ -41,14 +41,17 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() autoRowSize: Handsontable.GridSettings['autoRowSize'];
   @Input() autoWrapCol: Handsontable.GridSettings['autoWrapCol'];
   @Input() autoWrapRow: Handsontable.GridSettings['autoWrapRow'];
+  @Input() bindRowsWithHeaders: Handsontable.GridSettings['bindRowsWithHeaders'];
   @Input() cell: Handsontable.GridSettings['cell'];
   @Input() cells: Handsontable.GridSettings['cells'];
   @Input() checkedTemplate: Handsontable.GridSettings['checkedTemplate'];
   @Input() className: Handsontable.GridSettings['className'];
   @Input() colHeaders: Handsontable.GridSettings['colHeaders'];
+  @Input() collapsibleColumns: Handsontable.GridSettings['collapsibleColumns'];
   @Input() columnHeaderHeight: Handsontable.GridSettings['columnHeaderHeight'];
   @Input() columns: Handsontable.GridSettings['columns'];
   @Input() columnSorting: Handsontable.GridSettings['columnSorting'];
+  @Input() columnSummary: Handsontable.GridSettings['columnSummary'];
   @Input() colWidths: Handsontable.GridSettings['colWidths'];
   @Input() commentedCellClassName: Handsontable.GridSettings['commentedCellClassName'];
   @Input() comments: Handsontable.GridSettings['comments'];
@@ -67,20 +70,28 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() defaultDate: Handsontable.GridSettings['defaultDate'];
   @Input() disableVisualSelection: Handsontable.GridSettings['disableVisualSelection'];
   @Input() dragToScroll: Handsontable.GridSettings['dragToScroll'];
+  @Input() dropdownMenu: Handsontable.GridSettings['dropdownMenu'];
   @Input() editor: Handsontable.GridSettings['editor'];
   @Input() enterBeginsEditing: Handsontable.GridSettings['enterBeginsEditing'];
   @Input() enterMoves: Handsontable.GridSettings['enterMoves'];
   @Input() fillHandle: Handsontable.GridSettings['fillHandle'];
   @Input() filter: Handsontable.GridSettings['filter'];
   @Input() filteringCaseSensitive: Handsontable.GridSettings['filteringCaseSensitive'];
+  @Input() filters: Handsontable.GridSettings['filters'];
   @Input() fixedColumnsLeft: Handsontable.GridSettings['fixedColumnsLeft'];
   @Input() fixedRowsBottom: Handsontable.GridSettings['fixedRowsBottom'];
   @Input() fixedRowsTop: Handsontable.GridSettings['fixedRowsTop'];
+  @Input() formulas: Handsontable.GridSettings['formulas'];
   @Input() fragmentSelection: Handsontable.GridSettings['fragmentSelection'];
+  @Input() ganttChart: Handsontable.GridSettings['ganttChart'];
+  @Input() headerTooltips: Handsontable.GridSettings['headerTooltips'];
   @Input() height: Handsontable.GridSettings['height'];
+  @Input() hiddenColumns: Handsontable.GridSettings['hiddenColumns'];
+  @Input() hiddenRows: Handsontable.GridSettings['hiddenRows'];
   @Input() invalidCellClassName: Handsontable.GridSettings['invalidCellClassName'];
   @Input() label: Handsontable.GridSettings['label'];
   @Input() language: Handsontable.GridSettings['language'];
+  @Input() licenseKey: Handsontable.GridSettings['licenseKey'];
   @Input() manualColumnFreeze: Handsontable.GridSettings['manualColumnFreeze'];
   @Input() manualColumnMove: Handsontable.GridSettings['manualColumnMove'];
   @Input() manualColumnResize: Handsontable.GridSettings['manualColumnResize'];
@@ -93,6 +104,9 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() minRows: Handsontable.GridSettings['minRows'];
   @Input() minSpareCols: Handsontable.GridSettings['minSpareCols'];
   @Input() minSpareRows: Handsontable.GridSettings['minSpareRows'];
+  @Input() multiColumnSorting: Handsontable.GridSettings['multiColumnSorting'];
+  @Input() nestedHeaders: Handsontable.GridSettings['nestedHeaders'];
+  @Input() nestedRows: Handsontable.GridSettings['nestedRows'];
   @Input() noWordWrapClassName: Handsontable.GridSettings['noWordWrapClassName'];
   @Input() numericFormat: Handsontable.GridSettings['numericFormat'];
   @Input() observeChanges: Handsontable.GridSettings['observeChanges'];
@@ -123,6 +137,7 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() tabMoves: Handsontable.GridSettings['tabMoves'];
   @Input() title: Handsontable.GridSettings['title'];
   @Input() trimDropdown: Handsontable.GridSettings['trimDropdown'];
+  @Input() trimRows: Handsontable.GridSettings['nestedRows'];
   @Input() trimWhitespace: Handsontable.GridSettings['trimWhitespace'];
   @Input() type: Handsontable.GridSettings['type'];
   @Input() uncheckedTemplate: Handsontable.GridSettings['uncheckedTemplate'];
@@ -133,7 +148,9 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() visibleRows: Handsontable.GridSettings['visibleRows'];
   @Input() width: Handsontable.GridSettings['width'];
   @Input() wordWrap: Handsontable.GridSettings['wordWrap'];
+
   // handsontable hooks
+  @Input() afterAddChild: Handsontable.GridSettings['afterAddChild'];
   @Input() afterBeginEditing: Handsontable.GridSettings['afterBeginEditing'];
   @Input() afterCellMetaReset: Handsontable.GridSettings['afterCellMetaReset'];
   @Input() afterChange: Handsontable.GridSettings['afterChange'];
@@ -151,13 +168,20 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() afterCut: Handsontable.GridSettings['afterCut'];
   @Input() afterDeselect: Handsontable.GridSettings['afterDeselect'];
   @Input() afterDestroy: Handsontable.GridSettings['afterDestroy'];
+  @Input() afterDetachChild: Handsontable.GridSettings['afterDetachChild'];
   @Input() afterDocumentKeyDown: Handsontable.GridSettings['afterDocumentKeyDown'];
   @Input() afterDrawSelection: Handsontable.GridSettings['afterDrawSelection'];
+  @Input() afterDropdownMenuDefaultOptions: Handsontable.GridSettings['afterDropdownMenuDefaultOptions'];
+  @Input() afterDropdownMenuHide: Handsontable.GridSettings['afterDropdownMenuHide'];
+  @Input() afterDropdownMenuShow: Handsontable.GridSettings['afterDropdownMenuShow'];
+  @Input() afterFilter: Handsontable.GridSettings['afterFilter'];
   @Input() afterGetCellMeta: Handsontable.GridSettings['afterGetCellMeta'];
   @Input() afterGetColHeader: Handsontable.GridSettings['afterGetColHeader'];
   @Input() afterGetColumnHeaderRenderers: Handsontable.GridSettings['afterGetColumnHeaderRenderers'];
   @Input() afterGetRowHeader: Handsontable.GridSettings['afterGetRowHeader'];
   @Input() afterGetRowHeaderRenderers: Handsontable.GridSettings['afterGetRowHeaderRenderers'];
+  @Input() afterHideColumns: Handsontable.GridSettings['afterHideColumns'];
+  @Input() afterHideRows: Handsontable.GridSettings['afterHideRows'];
   @Input() afterInit: Handsontable.GridSettings['afterInit'];
   @Input() afterLanguageChange: Handsontable.GridSettings['afterLanguageChange'];
   @Input() afterListen: Handsontable.GridSettings['afterListen'];
@@ -176,6 +200,7 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() afterPaste: Handsontable.GridSettings['afterPaste'];
   @Input() afterPluginsInitialized: Handsontable.GridSettings['afterPluginsInitialized'];
   @Input() afterRedo: Handsontable.GridSettings['afterRedo'];
+  @Input() afterRefreshDimensions: Handsontable.GridSettings['afterRefreshDimensions'];
   @Input() afterRemoveCellMeta: Handsontable.GridSettings['afterRemoveCellMeta'];
   @Input() afterRemoveCol: Handsontable.GridSettings['afterRemoveCol'];
   @Input() afterRemoveRow: Handsontable.GridSettings['afterRemoveRow'];
@@ -192,13 +217,18 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() afterSetCellMeta: Handsontable.GridSettings['afterSetCellMeta'];
   @Input() afterSetDataAtCell: Handsontable.GridSettings['afterSetDataAtCell'];
   @Input() afterSetDataAtRowProp: Handsontable.GridSettings['afterSetDataAtRowProp'];
+  @Input() afterTrimRow: Handsontable.GridSettings['afterTrimRow'];
   @Input() afterUndo: Handsontable.GridSettings['afterUndo'];
+  @Input() afterUnhideColumns: Handsontable.GridSettings['afterUnhideColumns'];
+  @Input() afterUnhideRows: Handsontable.GridSettings['afterUnhideRows'];
   @Input() afterUnlisten: Handsontable.GridSettings['afterUnlisten'];
   @Input() afterUnmergeCells: Handsontable.GridSettings['afterUnmergeCells'];
+  @Input() afterUntrimRow: Handsontable.GridSettings['afterUntrimRow'];
   @Input() afterUpdateSettings: Handsontable.GridSettings['afterUpdateSettings'];
   @Input() afterValidate: Handsontable.GridSettings['afterValidate'];
   @Input() afterViewportColumnCalculatorOverride: Handsontable.GridSettings['afterViewportColumnCalculatorOverride'];
   @Input() afterViewportRowCalculatorOverride: Handsontable.GridSettings['afterViewportRowCalculatorOverride'];
+  @Input() beforeAddChild: Handsontable.GridSettings['beforeAddChild'];
   @Input() beforeAutofill: Handsontable.GridSettings['beforeAutofill'];
   @Input() beforeAutofillInsidePopulate: Handsontable.GridSettings['beforeAutofillInsidePopulate'];
   @Input() beforeCellAlignment: Handsontable.GridSettings['beforeCellAlignment'];
@@ -213,8 +243,14 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() beforeCreateCol: Handsontable.GridSettings['beforeCreateCol'];
   @Input() beforeCreateRow: Handsontable.GridSettings['beforeCreateRow'];
   @Input() beforeCut: Handsontable.GridSettings['beforeCut'];
+  @Input() beforeDetachChild: Handsontable.GridSettings['beforeDetachChild'];
   @Input() beforeDrawBorders: Handsontable.GridSettings['beforeDrawBorders'];
+  @Input() beforeDropdownMenuSetItems: Handsontable.GridSettings['beforeDropdownMenuSetItems'];
+  @Input() beforeDropdownMenuShow: Handsontable.GridSettings['beforeDropdownMenuShow'];
+  @Input() beforeFilter: Handsontable.GridSettings['beforeFilter'];
   @Input() beforeGetCellMeta: Handsontable.GridSettings['beforeGetCellMeta'];
+  @Input() beforeHideColumns: Handsontable.GridSettings['beforeHideColumns'];
+  @Input() beforeHideRows: Handsontable.GridSettings['beforeHideRows'];
   @Input() beforeInit: Handsontable.GridSettings['beforeInit'];
   @Input() beforeInitWalkontable: Handsontable.GridSettings['beforeInitWalkontable'];
   @Input() beforeKeyDown: Handsontable.GridSettings['beforeKeyDown'];
@@ -227,6 +263,7 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() beforeOnCellMouseUp: Handsontable.GridSettings['beforeOnCellMouseUp'];
   @Input() beforePaste: Handsontable.GridSettings['beforePaste'];
   @Input() beforeRedo: Handsontable.GridSettings['beforeRedo'];
+  @Input() beforeRefreshDimensions: Handsontable.GridSettings['beforeRefreshDimensions'];
   @Input() beforeRemoveCellClassNames: Handsontable.GridSettings['beforeRemoveCellClassNames'];
   @Input() beforeRemoveCellMeta: Handsontable.GridSettings['beforeRemoveCellMeta'];
   @Input() beforeRemoveCol: Handsontable.GridSettings['beforeRemoveCol'];
@@ -240,11 +277,17 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() beforeSetRangeStartOnly: Handsontable.GridSettings['beforeSetRangeStartOnly'];
   @Input() beforeStretchingColumnWidth: Handsontable.GridSettings['beforeStretchingColumnWidth'];
   @Input() beforeTouchScroll: Handsontable.GridSettings['beforeTouchScroll'];
+  @Input() beforeTrimRow: Handsontable.GridSettings['beforeTrimRow'];
   @Input() beforeUndo: Handsontable.GridSettings['beforeUndo'];
+  @Input() beforeUnhideColumns: Handsontable.GridSettings['beforeUnhideColumns'];
+  @Input() beforeUnhideRows: Handsontable.GridSettings['beforeUnhideRows'];
   @Input() beforeUnmergeCells: Handsontable.GridSettings['beforeUnmergeCells'];
+  @Input() beforeUntrimRow: Handsontable.GridSettings['beforeUntrimRow'];
   @Input() beforeValidate: Handsontable.GridSettings['beforeValidate'];
   @Input() beforeValueRender: Handsontable.GridSettings['beforeValueRender'];
   @Input() construct: Handsontable.GridSettings['construct'];
+  @Input() hiddenColumn: Handsontable.GridSettings['hiddenColumn'];
+  @Input() hiddenRow: Handsontable.GridSettings['hiddenRow'];
   @Input() init: Handsontable.GridSettings['init'];
   @Input() modifyAutofillRange: Handsontable.GridSettings['modifyAutofillRange'];
   @Input() modifyCol: Handsontable.GridSettings['modifyCol'];
@@ -259,7 +302,6 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() modifyRowHeader: Handsontable.GridSettings['modifyRowHeader'];
   @Input() modifyRowHeaderWidth: Handsontable.GridSettings['modifyRowHeaderWidth'];
   @Input() modifyRowHeight: Handsontable.GridSettings['modifyRowHeight'];
-  @Input() modifyRowSourceData: Handsontable.GridSettings['modifyRowSourceData'];
   @Input() modifyTransformEnd: Handsontable.GridSettings['modifyTransformEnd'];
   @Input() modifyTransformStart: Handsontable.GridSettings['modifyTransformStart'];
   @Input() persistentStateLoad: Handsontable.GridSettings['persistentStateLoad'];
@@ -268,38 +310,6 @@ export class HotTableComponent implements AfterContentInit, OnChanges, OnDestroy
   @Input() skipLengthCache: Handsontable.GridSettings['skipLengthCache'];
   @Input() unmodifyCol: Handsontable.GridSettings['unmodifyCol'];
   @Input() unmodifyRow: Handsontable.GridSettings['unmodifyRow'];
-  // handsontable-pro options
-  @Input() bindRowsWithHeaders: Handsontable.GridSettings['bindRowsWithHeaders'];
-  @Input() collapsibleColumns: Handsontable.GridSettings['collapsibleColumns'];
-  @Input() columnSummary: Handsontable.GridSettings['columnSummary'];
-  @Input() dropdownMenu: Handsontable.GridSettings['dropdownMenu'];
-  @Input() filters: Handsontable.GridSettings['filters'];
-  @Input() formulas: Handsontable.GridSettings['formulas'];
-  @Input() ganttChart: Handsontable.GridSettings['ganttChart'];
-  @Input() headerTooltips: Handsontable.GridSettings['headerTooltips'];
-  @Input() hiddenColumns: Handsontable.GridSettings['hiddenColumns'];
-  @Input() hiddenRows: Handsontable.GridSettings['hiddenRows'];
-  @Input() licenseKey: Handsontable.GridSettings['licenseKey'];
-  @Input() multiColumnSorting: Handsontable.GridSettings['multiColumnSorting'];
-  @Input() nestedHeaders: Handsontable.GridSettings['nestedHeaders'];
-  @Input() nestedRows: Handsontable.GridSettings['nestedRows'];
-  @Input() trimRows: Handsontable.GridSettings['nestedRows'];
-  // handsontable-pro hooks
-  @Input() afterAddChild: Handsontable.GridSettings['afterAddChild'];
-  @Input() afterDetachChild: Handsontable.GridSettings['afterDetachChild'];
-  @Input() afterDropdownMenuDefaultOptions: Handsontable.GridSettings['afterDropdownMenuDefaultOptions'];
-  @Input() afterDropdownMenuHide: Handsontable.GridSettings['afterDropdownMenuHide'];
-  @Input() afterDropdownMenuShow: Handsontable.GridSettings['afterDropdownMenuShow'];
-  @Input() afterFilter: Handsontable.GridSettings['afterFilter'];
-  @Input() afterTrimRow: Handsontable.GridSettings['afterTrimRow'];
-  @Input() afterUntrimRow: Handsontable.GridSettings['afterUntrimRow'];
-  @Input() beforeAddChild: Handsontable.GridSettings['beforeAddChild'];
-  @Input() beforeDetachChild: Handsontable.GridSettings['beforeDetachChild'];
-  @Input() beforeDropdownMenuSetItems: Handsontable.GridSettings['beforeDropdownMenuSetItems'];
-  @Input() beforeDropdownMenuShow: Handsontable.GridSettings['beforeDropdownMenuShow'];
-  @Input() beforeFilter: Handsontable.GridSettings['beforeFilter'];
-  @Input() hiddenColumn: Handsontable.GridSettings['hiddenColumn'];
-  @Input() hiddenRow: Handsontable.GridSettings['hiddenRow'];
 
   constructor(
     private _ngZone: NgZone,
